Выделен отдельный компонент EnhancedTableHead
